+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `preanniversary-phrases` as a builtin dictionary and default search dictionary
+- `--show-dictionary` option for `grascii search`
 
 ### Changed